Refactor Firebase integration and add type definitions for meals and recipes
