+++ package.json
@@ -12,6 +12,7 @@
     "@react-native-async-storage/async-storage": "^1.24.0",
     "@react-native-firebase/firestore": "^21.12.0",
     "@react-native-community/datetimepicker": "^8.2.0",
+    "@react-native-firebase/auth": "^21.12.0",
     "expo": "^52.0.39",
     "expo-constants": "~17.0.8",
     "expo-linking": "~7.0.5",
