+++ tsconfig.json
@@ -7,6 +7,8 @@
     "paths": {
       "@/*": ["./*"],
       "@firebase/auth": ["./node_modules/@firebase/auth/dist/index.rn.d.ts"]
-	    }
+    },
+    "noImplicitAny": true,
+    "strictNullChecks": true
   }
 }
