+++ types.ts
@@ -1,3 +1,5 @@
+import { Timestamp } from "firebase/firestore";
+
 export interface RecipeData {
   id?: string;
   name: string;
@@ -6,4 +8,42 @@ export interface RecipeData {
   photoUrl?: string;
   createdAt?: any;
   isFavorite?: boolean;
+
+}
+
+export interface Meal {
+  date: Timestamp;
+  type: string;
+  image?: string;
+  ingredients: string[];
+  analyzed: boolean;
+}
+
+export interface mealsFromDB {
+  id: string;
+  date: Timestamp;
+  type: string;
+  image?: string;
+  ingredients: string[];
+  analyzed: boolean;
+  nutrition?: Nutrition;
+}
+
+export interface Nutrition {
+  calories: number;
+  totalNutrients: {
+    [key: string]: {
+      label: string;
+      quantity: number;
+      unit: string;
+    };
+  };
+  totalDaily: {
+    [key: string]: {
+      label: string;
+      quantity: number;
+      unit: string;
+    };
+  };
+
 }
